[MIG] sale_timesheet_invoice_description: Migration to 16.0
* adapted code to version 16.0 changes; for details see below
* added contributors
* updated i18n: sale_timesheet_invoice_description.pot, de.po
* run `pre-commit`

Linking timesheets to invoice lines: Respecting both the specified (in
context) start and end time, the same way as it is done for linking them
to the invoice itself.

Reworked preventing unbalanced invoice: Invoices must not and cannot be
unbalanced. This is now (in 16.0) done automatically. I.e., the
intermediate invoices, created while splitting lines, will no longer be
possibly unbalanced.